♻️ Refactors boss classes and removes documentation files
Moves boss entity implementations from inline definitions to separate class files following the AbstractEntity pattern established for other enemy types.

Removes outdated refactoring documentation files that are no longer needed after completing the entity system reorganization.

Updates boss collision detection to use proper entity coordinates and adds missing properties to base Boss class for consistency with the new architecture.

Fixes ScrollingBackground component by adding proper imports and exports that were previously missing.
