+++ src/components/GameCanvas.tsx
@@ -2,14 +2,13 @@ import React, { useEffect, useRef, useState } from 'react';
 import Phaser from 'phaser';
 import GameUI from "./GameUI";
 
-import Enemy from "../game/entities/Enemy";
 import EnemyTypeA from "../game/entities/EnemyTypeA";
 import EnemyTypeB from "../game/entities/EnemyTypeB";
 import EnemyTypeC from "../game/entities/EnemyTypeC";
-import Player from "../game/entities/Player";
-import SpawnSystem from "../game/systems/SpawnSystem";
-import WaveManager from "../game/systems/WaveManager";
-import LifeSystem from "../game/systems/LifeSystem";
+import BossTypeA from "../game/entities/BossTypeA";
+import BossTypeB from "../game/entities/BossTypeB";
+import BossTypeC from "../game/entities/BossTypeC";
+import ScrollingBackground from "./ScrollingBackground";
 
 // Importar as imagens do player
 import playerFrame1 from '../assets/images/player/player-frame-1.png';
@@ -271,1170 +270,6 @@ const GameCanvas: React.FC<GameCanvasProps> = ({ backgroundSpeed = .75, onNaviga
       }
     }
 
-
-    // =============================================
-    // ENEMY TYPE A - AGORA EM ARQUIVO SEPARADO
-    // =============================================
-    // A classe EnemyTypeA foi movida para src/game/entities/EnemyTypeA.ts
-    // e agora herda corretamente de AbstractEntity com todas as funcionalidades otimizadas.
-
-    // =============================================
-    // ENEMY TYPE B - AGORA EM ARQUIVO SEPARADO
-    // =============================================
-    // A classe EnemyTypeB foi movida para src/game/entities/EnemyTypeB.ts
-    // e agora herda corretamente de AbstractEntity com todas as funcionalidades otimizadas.
-
-    // =============================================
-    // ENEMY TYPE C - AGORA EM ARQUIVO SEPARADO
-    // =============================================
-    // A classe EnemyTypeC foi movida para src/game/entities/EnemyTypeC.ts
-    // e agora herda corretamente de AbstractEntity com todas as funcionalidades otimizadas.
-
-    // =============================================
-    // BOSS TYPE A
-    // =============================================
-    class BossTypeA extends Enemy {
-      // Estados do chefe
-      private state: 'entering' | 'moving_to_position' | 'shooting' | 'moving' | 'leaving' = 'entering';
-      
-      // Sistema de disparo
-      private fireTimer = 0;
-      private fireRate = 1000; // 1 segundo entre rajadas
-      private burstTimer = 0;
-      private burstRate = 200; // 200ms entre tiros da mesma rajada
-      private currentBurstShots = 0;
-      private isInBurst = false;
-      private burstCount = 0;
-      private maxBursts = 3; // 3 rajadas por posição
-      
-      // Sistema de movimento
-      private moveSpeed = 120; // Mais lento que inimigo tipo A (150)
-      private targetX = 0;
-      private targetY = 0;
-      private positionsVisited = 0;
-      
-      // Sistema de dano visual
-      private isFlashing = false;
-      
-      constructor(scene: Phaser.Scene, x: number, y: number) {
-        super(scene, x, y, 'boss-A-frame-1');
-        
-        // Stats específicos do Boss Type A
-        this.hp = 50;
-        this.maxHp = 100;
-        this.speed = 120;
-        this.maxSpeed = 200;
-        this.shotsPerBurst = 3; // 3 tiros por rajada
-        this.maxShotsPerBurst = 7;
-        this.totalShots = 30; // Será ajustado pela onda
-        this.maxTotalShots = 100;
-        this.movementsBeforeLeaving = 999; // Nunca vai embora
-        this.maxMovements = 999;
-        
-        // Configurar sprite
-        this.sprite.setScale(1.5); // Maior que inimigos normais
-        this.sprite.setData('enemyType', 'BossA');
-        this.sprite.setData('hp', this.hp);
-        this.sprite.setData('maxHp', this.maxHp);
-        
-        // Definir primeira posição aleatória na metade superior
-        this.setRandomTarget();
-        
-        console.log(`Boss Type A criado com ${this.hp} HP`);
-      }
-      
-      // Sobrescrever método para calcular pontos baseado na onda
-      adjustForWave(wave: number) {
-        // Pontuação baseada na onda
-        this.points = 1000 * wave;
-        
-        // Aumentar shots per burst baseado na onda
-        this.shotsPerBurst = Math.min(this.maxShotsPerBurst, 3 + Math.floor(wave / 2));
-        
-        // Aumentar total shots baseado na onda
-        this.totalShots = Math.min(this.maxTotalShots, 30 + (wave * 10));
-        
-        // Aumentar velocidade ligeiramente baseado na onda
-        this.speed = Math.min(this.maxSpeed, 120 + (wave * 5));
-        this.moveSpeed = this.speed;
-        
-        console.log(`Boss ajustado para onda ${wave}: ${this.shotsPerBurst} tiros/rajada, ${this.points} pontos`);
-      }
-      
-      private setRandomTarget() {
-        // Posição aleatória na metade superior da tela
-        const preferredX = Phaser.Math.Between(200, 600);
-        const preferredY = Phaser.Math.Between(100, 300);
-        
-        // Boss não precisa verificar posições ocupadas (ele tem prioridade)
-        this.targetX = preferredX;
-        this.targetY = preferredY;
-      }
-      
-      move() {
-        const deltaTime = this.scene.game.loop.delta / 1000;
-        
-        // Calcular direção para o alvo
-        const directionX = this.targetX - this.sprite.x;
-        const directionY = this.targetY - this.sprite.y;
-        const distance = Math.sqrt(directionX * directionX + directionY * directionY);
-        
-        // Se chegou perto o suficiente do alvo
-        if (distance < 15) {
-          this.onReachedTarget();
-          return;
-        }
-        
-        // Normalizar direção e aplicar velocidade
-        const normalizedX = directionX / distance;
-        const normalizedY = directionY / distance;
-        
-        this.sprite.x += normalizedX * this.moveSpeed * deltaTime;
-        this.sprite.y += normalizedY * this.moveSpeed * deltaTime;
-        
-        // Atualizar animação de movimento
-        this.updateMovementAnimation(normalizedX);
-      }
-      
-      private onReachedTarget() {
-        // Chegou na posição, começar a atirar
-        this.state = 'shooting';
-        this.resetShooting();
-        this.positionsVisited++;
-      }
-      
-      private resetShooting() {
-        this.burstCount = 0;
-        this.shotsFired = 0;
-        this.fireTimer = this.scene.time.now;
-        
-        // Só mudar textura se não estiver em flash E se a textura for diferente
-        if (!this.isFlashing && this.sprite.texture.key !== 'boss-A-frame-1') {
-          this.sprite.setTexture('boss-A-frame-1'); // Frame de repouso
-        }
-      }
-      
-      shoot() {
-        // Reproduzir som do tiro do boss (volume maior para diferenciá-lo dos inimigos)
-        this.scene.sound.play('enemy-shoot', { volume: 0.4 });
-        
-        // Obter referência ao grupo de projéteis inimigos da scene
-        const gameScene = this.scene as GameScene;
-        if (!gameScene.enemyBullets) return;
-        
-        // Tentar criar projétil do boss partindo do centro, sempre em linha reta para baixo
-        let bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + 40, 'boss-fire');
-        
-        // Sistema de fallback para garantir que o projétil seja criado
-        if (!bullet) {
-          // Limpar projéteis que saíram da tela para liberar espaço no grupo
-          gameScene.enemyBullets.children.entries.forEach((child: Phaser.GameObjects.GameObject) => {
-            const sprite = child as Phaser.GameObjects.Sprite;
-            if (sprite.active && (sprite.y > 650 || sprite.y < -50 || sprite.x < -50 || sprite.x > 850)) {
-              sprite.setActive(false).setVisible(false);
-            }
-          });
-          
-          // Tentar novamente após limpeza
-          bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + 40, 'boss-fire');
-        }
-        
-        if (bullet) {
-          bullet.setActive(true);
-          bullet.setVisible(true);
-          bullet.setScale(1.2); // Projéteis do boss são maiores
-          bullet.setData('speed', 350); // Velocidade do projétil apenas para baixo
-          bullet.setData('damage', 2); // Boss causa mais dano
-          bullet.setData('isEnemyBullet', true);
-          bullet.setData('isBossBullet', true); // Identificar como projétil de boss
-          // Garantir que não tem flags de movimento angular
-          bullet.setData('isAngled', false);
-          bullet.setData('isDirected', false);
-        } else {
-          console.warn('Boss A não conseguiu criar projétil');
-        }
-        
-        this.shotsFired++;
-      }
-      
-      private updateMovementAnimation(directionX: number) {
-        // Atualizar frame baseado na direção do movimento (apenas se não estiver em flash)
-        if (this.isFlashing) return;
-        
-        if (Math.abs(directionX) > 0.1) {
-          // Está se movendo horizontalmente
-          const frame = (this.scene.time.now % 600 < 300) ? 2 : 3;
-          const newTexture = `boss-A-frame-${frame}`;
-          
-          // Só mudar textura se for diferente da atual
-          if (this.sprite.texture.key !== newTexture) {
-            this.sprite.setTexture(newTexture);
-          }
-          
-          // Espelhar sprite baseado na direção
-          this.sprite.setFlipX(directionX > 0);
-        } else {
-          // Movimento apenas vertical ou parado
-          if (this.sprite.texture.key !== 'boss-A-frame-1') {
-            this.sprite.setTexture('boss-A-frame-1');
-          }
-          this.sprite.setFlipX(false);
-        }
-      }
-      
-      // Sobrescrever método takeDamage para adicionar efeito de flash
-      takeDamage(damage: number) {
-        this.hp -= damage;
-        console.log(`Boss levou ${damage} de dano. HP restante: ${this.hp}/${this.maxHp}`);
-        
-        // Ativar efeito de flash
-        this.activateFlash();
-        
-        if (this.hp <= 0) {
-          this.destroy();
-        }
-      }
-      
-      private activateFlash() {
-        if (this.isFlashing) return; // Evitar múltiplos flashes simultâneos
-        
-        this.isFlashing = true;
-        console.log('Boss piscando!'); // Debug temporário
-        
-        // Mudar para cor branca (flash) - usando tint mais intenso
-        this.sprite.setTintFill(0xffffff);
-        
-        // Voltar à cor normal após 250ms (aumentado ainda mais para melhor visibilidade)
-        this.scene.time.delayedCall(50, () => {
-          if (this.sprite && !this.isDestroyed) {
-            this.sprite.clearTint();
-            console.log('Flash do boss terminado'); // Debug temporário
-          }
-          this.isFlashing = false;
-        });
-      }
-      
-      // Sobrescrever método destroy para múltiplas animações de destruição
-      destroy() {
-        if (this.isDestroyed) return;
-        
-        this.isDestroyed = true;
-        this.isKilledByPlayer = true;
-        
-        // Reproduzir som de morte do boss
-        this.scene.sound.play('boss-kill', { volume: 0.5 });
-        
-        console.log(`Boss Type A destruído! Pontuação: ${this.points}`);
-        
-        // Criar 5 animações de destruição: centro + 4 cantos
-        const centerX = this.sprite.x;
-        const centerY = this.sprite.y;
-        const offset = 40; // Distância dos cantos em relação ao centro
-        
-        // Animação central
-        new DeathAnimation(this.scene, centerX, centerY);
-        
-        // Animações nos cantos (com pequeno delay para efeito visual)
-        this.scene.time.delayedCall(100, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY - offset); // Superior esquerdo
-        });
-        
-        this.scene.time.delayedCall(200, () => {
-          new DeathAnimation(this.scene, centerX + offset, centerY - offset); // Superior direito
-        });
-        
-        this.scene.time.delayedCall(300, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY + offset); // Inferior esquerdo
-        });
-        
-        this.scene.time.delayedCall(400, () => {
-          new DeathAnimation(this.scene, centerX + offset, centerY + offset, () => {
-            // Callback final após todas as animações - iniciar transição de onda
-            const gameScene = this.scene as GameScene;
-            gameScene.onBossDefeated(this.points);
-          });
-        });
-        
-        // Remover sprite do boss
-        this.sprite.destroy();
-      }
-      
-      update() {
-        super.update();
-        if (this.isDestroyed) return;
-        
-        const currentTime = this.scene.time.now;
-        
-        // Máquina de estados
-        switch (this.state) {
-          case 'entering':
-            // Verificar se entrou na tela o suficiente
-            if (this.sprite.y > 80) {
-              this.state = 'moving_to_position';
-            } else {
-              // Continuar entrando (movimento para baixo)
-              this.sprite.y += this.moveSpeed * (this.scene.game.loop.delta / 1000);
-            }
-            break;
-            
-          case 'moving_to_position':
-          case 'moving':
-            this.move();
-            break;
-            
-          case 'shooting':
-            // Executar comportamento de tiro
-            if (!this.isInBurst) {
-              if (currentTime - this.fireTimer > this.fireRate && this.burstCount < this.maxBursts) {
-                this.isInBurst = true;
-                this.currentBurstShots = 0;
-                this.burstTimer = currentTime;
-              } else if (this.burstCount >= this.maxBursts) {
-                // Completou os tiros nesta posição, mover para próxima
-                this.onCompletedShooting();
-              }
-            } else {
-              // Executando rajada
-              if (currentTime - this.burstTimer > this.burstRate && this.currentBurstShots < this.shotsPerBurst) {
-                this.shoot();
-                this.currentBurstShots++;
-                this.burstTimer = currentTime;
-                
-                // Verificar se rajada terminou
-                if (this.currentBurstShots >= this.shotsPerBurst) {
-                  this.isInBurst = false;
-                  this.burstCount++;
-                  this.fireTimer = currentTime;
-                }
-              }
-            }
-            break;
-        }
-      }
-      
-      private onCompletedShooting() {
-        // Boss nunca vai embora, sempre move para nova posição
-        this.setRandomTarget();
-        this.state = 'moving';
-      }
-    }
-
-    // =============================================
-    // BOSS TYPE B
-    // =============================================
-    class BossTypeB extends Enemy {
-      // Estados do chefe (baseado no Enemy Type B)
-      private state: 'entering' | 'initial_shooting' | 'moving_to_position' | 'position_shooting' | 'moving' = 'entering';
-      
-      // Sistema de disparo
-      private fireTimer = 0;
-      private fireRate = 1500; // 1.5 segundos entre rajadas
-      private burstTimer = 0;
-      private burstRate = 400; // 400ms entre tiros da mesma rajada
-      private currentBurstShots = 0;
-      private isInBurst = false;
-      private burstCount = 0;
-      private maxBursts = 1; // 1 rajada por posição
-      
-      // Sistema de movimento
-      private moveSpeed = 100; // Mais lento que o Enemy Type B (150)
-      private targetX = 0;
-      private targetY = 0;
-      public maxMovements = 3; // Mover 3 vezes
-      private hasCompletedInitialShooting = false;
-      private positions: { x: number; y: number }[] = [];
-      private movementCountB = 0;
-      
-      // Sistema de dano visual
-      private isFlashing = false;
-      
-      constructor(scene: Phaser.Scene, x: number, y: number) {
-        super(scene, x, y, 'boss-B-frame-1');
-        
-        // Stats específicos do Boss Type B
-        this.hp = 60;
-        this.maxHp = 120;
-        this.speed = 100;
-        this.maxSpeed = 150;
-        this.shotsPerBurst = 3; // 3 tiros por rajada (cada tiro = 3 projéteis)
-        this.maxShotsPerBurst = 6;
-        this.totalShots = 30; // Será ajustado pela onda
-        this.maxTotalShots = 100;
-        this.movementsBeforeLeaving = 999; // Nunca vai embora
-        this.maxMovements = 999;
-        
-        // Configurar sprite
-        this.sprite.setScale(1.5); // Maior que inimigos normais
-        this.sprite.setData('enemyType', 'BossB');
-        this.sprite.setData('hp', this.hp);
-        this.sprite.setData('maxHp', this.maxHp);
-        
-        // Gerar 3 posições aleatórias que o boss visitará
-        this.generateRandomPositions();
-        
-        // Começar no estado de entrada
-        this.state = 'entering';
-        
-        console.log(`Boss Type B criado com ${this.hp} HP`);
-      }
-      
-      // Sobrescrever método para calcular pontos baseado na onda
-      adjustForWave(wave: number) {
-        // Pontuação baseada na onda
-        this.points = 3000 * wave;
-        
-        // Aumentar shots per burst baseado na onda
-        this.shotsPerBurst = Math.min(this.maxShotsPerBurst, 3 + Math.floor(wave / 2));
-        
-        // Aumentar total shots baseado na onda
-        this.totalShots = Math.min(this.maxTotalShots, 30 + (wave * 10));
-        
-        // Aumentar velocidade ligeiramente baseado na onda
-        this.speed = Math.min(this.maxSpeed, 100 + (wave * 3));
-        this.moveSpeed = this.speed;
-        
-        console.log(`Boss B ajustado para onda ${wave}: ${this.shotsPerBurst} tiros/rajada, ${this.points} pontos`);
-      }
-      
-      private generateRandomPositions() {
-        // Gerar 3 posições aleatórias na metade superior da tela (similar ao Enemy B)
-        const basePositions = [];
-        
-        for (let i = 0; i < 3; i++) {
-          let x: number, y: number;
-          
-          switch (i) {
-            case 0:
-              // Primeira posição: lado esquerdo a centro-esquerdo
-              x = Phaser.Math.Between(150, 350);
-              y = Phaser.Math.Between(120, 200);
-              break;
-            case 1:
-              // Segunda posição: centro-direita a direita
-              x = Phaser.Math.Between(450, 650);
-              y = Phaser.Math.Between(140, 220);
-              break;
-            case 2:
-              // Terceira posição: centro da tela
-              x = Phaser.Math.Between(300, 500);
-              y = Phaser.Math.Between(180, 260);
-              break;
-            default:
-              x = Phaser.Math.Between(200, 600);
-              y = Phaser.Math.Between(120, 250);
-          }
-          
-          basePositions.push({ x, y });
-        }
-        
-        // Boss não precisa verificar posições ocupadas (ele tem prioridade)
-        this.positions = basePositions;
-      }
-      
-      move() {
-        const deltaTime = this.scene.game.loop.delta / 1000;
-        
-        // Calcular direção para o alvo
-        const directionX = this.targetX - this.sprite.x;
-        const directionY = this.targetY - this.sprite.y;
-        const distance = Math.sqrt(directionX * directionX + directionY * directionY);
-        
-        // Se chegou perto o suficiente do alvo
-        if (distance < 15) {
-          this.onReachedTarget();
-          return;
-        }
-        
-        // Normalizar direção e aplicar velocidade
-        const normalizedX = directionX / distance;
-        const normalizedY = directionY / distance;
-        
-        this.sprite.x += normalizedX * this.moveSpeed * deltaTime;
-        this.sprite.y += normalizedY * this.moveSpeed * deltaTime;
-        
-        // Atualizar animação de movimento
-        this.updateMovementAnimation(normalizedX);
-      }
-      
-      private onReachedTarget() {
-        // Chegou na posição, mudar para modo de tiro
-        this.state = 'position_shooting';
-        this.resetShooting();
-      }
-      
-      private setNextTarget() {
-        if (this.movementCountB < this.maxMovements && this.movementCountB < this.positions.length) {
-          // Usar próxima posição aleatória gerada
-          const targetPosition = this.positions[this.movementCountB];
-          this.targetX = targetPosition.x;
-          this.targetY = targetPosition.y;
-          this.movementCountB++;
-        } else {
-          // Completou todos os movimentos, reiniciar ciclo
-          this.movementCountB = 0;
-          this.generateRandomPositions(); // Gerar novas posições
-          const targetPosition = this.positions[this.movementCountB];
-          this.targetX = targetPosition.x;
-          this.targetY = targetPosition.y;
-          this.movementCountB++;
-        }
-      }
-      
-      shoot() {
-        // Reproduzir som do tiro do boss (volume maior para diferenciá-lo dos inimigos)
-        this.scene.sound.play('enemy-shoot', { volume: 0.4 });
-        
-        // Obter referência ao grupo de projéteis inimigos da scene
-        const gameScene = this.scene as GameScene;
-        if (!gameScene.enemyBullets) return;
-        
-        // Disparar três projéteis: central, 45° esquerda, 45° direita
-        this.shootTriple();
-        
-        this.shotsFired++;
-      }
-      
-      private shootTriple() {
-        const gameScene = this.scene as GameScene;
-        if (!gameScene.enemyBullets) return;
-        
-        // Projétil central (direção direta para baixo)
-        this.shootAngledBoss(0);
-        
-        // Projétil para esquerda (45°)
-        this.shootAngledBoss(-45);
-        
-        // Projétil para direita (45°)
-        this.shootAngledBoss(45);
-      }
-      
-      private shootAngledBoss(angleDegrees: number) {
-        const gameScene = this.scene as GameScene;
-        if (!gameScene.enemyBullets) return;
-        
-        // Converter ângulo para radianos
-        const angleRadians = (angleDegrees * Math.PI) / 180;
-        
-        // Calcular posição de origem na borda inferior central do sprite
-        // Considerando que o sprite tem scale 1.5 e altura aproximada de 64 pixels
-        const spriteHeight = 64 * this.sprite.scaleY; // 64 * 1.5 = 96 pixels
-        const offsetY = spriteHeight / 2; // Metade da altura para chegar à borda inferior
-        
-        // Tentar criar projétil do boss na borda inferior central
-        // Com sistema de fallback para garantir que o projétil seja criado
-        let bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + offsetY, 'boss-fire');
-        
-        // Se não conseguiu obter um projétil, forçar limpeza de projéteis inativos e tentar novamente
-        if (!bullet) {
-          // Limpar projéteis que saíram da tela para liberar espaço no grupo
-          gameScene.enemyBullets.children.entries.forEach((child: Phaser.GameObjects.GameObject) => {
-            const sprite = child as Phaser.GameObjects.Sprite;
-            if (sprite.active && (sprite.y > 650 || sprite.y < -50 || sprite.x < -50 || sprite.x > 850)) {
-              sprite.setActive(false).setVisible(false);
-            }
-          });
-          
-          // Tentar novamente após limpeza
-          bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + offsetY, 'boss-fire');
-        }
-        
-        if (bullet) {
-          bullet.setActive(true);
-          bullet.setVisible(true);
-          bullet.setScale(1.2); // Projéteis do boss são maiores
-          
-          // Calcular velocidade baseada no ângulo
-          const speed = 320;
-          let velocityX, velocityY;
-          
-          if (angleDegrees === 0) {
-            // Projétil central - movimento direto para baixo
-            velocityX = 0;
-            velocityY = speed;
-          } else {
-            // Projéteis angulares
-            velocityX = Math.sin(angleRadians) * speed;
-            velocityY = Math.cos(angleRadians) * speed;
-          }
-          
-          bullet.setData('velocityX', velocityX);
-          bullet.setData('velocityY', velocityY);
-          bullet.setData('damage', 2); // Boss causa mais dano
-          bullet.setData('isEnemyBullet', true);
-          bullet.setData('isBossBullet', true);
-          bullet.setData('isAngled', angleDegrees !== 0); // Marcar projéteis angulares
-        } else {
-          // Se ainda assim não conseguir criar o projétil, registrar o problema
-          console.warn(`Boss B não conseguiu criar projétil em ângulo ${angleDegrees}°`);
-        }
-      }
-      
-      private updateMovementAnimation(directionX: number) {
-        // Atualizar frame baseado na direção do movimento (apenas se não estiver em flash)
-        if (this.isFlashing) return;
-        
-        if (Math.abs(directionX) > 0.1) {
-          // Está se movendo horizontalmente - alternar entre frames 2 e 3
-          const frame = (this.scene.time.now % 600 < 300) ? 2 : 3;
-          const newTexture = `boss-B-frame-${frame}`;
-          
-          // Só mudar textura se for diferente da atual
-          if (this.sprite.texture.key !== newTexture) {
-            this.sprite.setTexture(newTexture);
-          }
-          
-          // Espelhar sprite baseado na direção
-          this.sprite.setFlipX(directionX > 0);
-        } else {
-          // Movimento apenas vertical ou parado - usar frame 1
-          if (this.sprite.texture.key !== 'boss-B-frame-1') {
-            this.sprite.setTexture('boss-B-frame-1');
-          }
-          this.sprite.setFlipX(false);
-        }
-      }
-      
-      private resetShooting() {
-        this.burstCount = 0;
-        this.shotsFired = 0;
-        this.fireTimer = this.scene.time.now;
-        
-        // Só mudar textura se não estiver em flash
-        if (!this.isFlashing && this.sprite.texture.key !== 'boss-B-frame-1') {
-          this.sprite.setTexture('boss-B-frame-1'); // Frame de repouso
-        }
-      }
-      
-      // Sobrescrever método takeDamage para adicionar efeito de flash
-      takeDamage(damage: number) {
-        this.hp -= damage;
-        console.log(`Boss B levou ${damage} de dano. HP restante: ${this.hp}/${this.maxHp}`);
-        
-        // Ativar efeito de flash
-        this.activateFlash();
-        
-        if (this.hp <= 0) {
-          this.destroy();
-        }
-      }
-      
-      private activateFlash() {
-        if (this.isFlashing) return; // Evitar múltiplos flashes simultâneos
-        
-        this.isFlashing = true;
-        console.log('Boss B piscando!');
-        
-        // Mudar para cor branca (flash)
-        this.sprite.setTintFill(0xffffff);
-        
-        // Voltar à cor normal após 250ms
-        this.scene.time.delayedCall(50, () => {
-          if (this.sprite && !this.isDestroyed) {
-            this.sprite.clearTint();
-            console.log('Flash do boss B terminado');
-          }
-          this.isFlashing = false;
-        });
-      }
-      
-      // Sobrescrever método destroy para múltiplas animações de destruição
-      destroy() {
-        if (this.isDestroyed) return;
-        
-        this.isDestroyed = true;
-        this.isKilledByPlayer = true;
-        
-        // Reproduzir som de morte do boss
-        this.scene.sound.play('boss-kill', { volume: 0.5 });
-        
-        console.log(`Boss Type B destruído! Pontuação: ${this.points}`);
-        
-        // Criar 7 animações de destruição: centro + 6 ao redor
-        const centerX = this.sprite.x;
-        const centerY = this.sprite.y;
-        const offset = 50; // Distância maior para boss maior
-        
-        // Animação central
-        new DeathAnimation(this.scene, centerX, centerY);
-        
-        // Animações ao redor (com pequeno delay para efeito visual)
-        this.scene.time.delayedCall(80, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY - offset); // Superior esquerdo
-          new DeathAnimation(this.scene, centerX + offset, centerY - offset); // Superior direito
-        });
-        
-        this.scene.time.delayedCall(160, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY); // Esquerda
-          new DeathAnimation(this.scene, centerX + offset, centerY); // Direita
-        });
-        
-        this.scene.time.delayedCall(240, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY + offset); // Inferior esquerdo
-        });
-        
-        this.scene.time.delayedCall(320, () => {
-          new DeathAnimation(this.scene, centerX + offset, centerY + offset, () => {
-            // Callback final após todas as animações - iniciar transição de onda
-            const gameScene = this.scene as GameScene;
-            gameScene.onBossDefeated(this.points);
-          });
-        });
-        
-        // Remover sprite do boss
-        this.sprite.destroy();
-      }
-      
-      update() {
-        super.update();
-        if (this.isDestroyed) return;
-        
-        const currentTime = this.scene.time.now;
-        
-        // Máquina de estados (baseada no Enemy Type B)
-        switch (this.state) {
-          case 'entering':
-            // Verificar se entrou na tela
-            if (this.sprite.y > 80) {
-              this.state = 'initial_shooting';
-              this.resetShooting();
-            } else {
-              // Continuar entrando (movimento para baixo)
-              this.sprite.y += 100 * (this.scene.game.loop.delta / 1000);
-            }
-            break;
-            
-          case 'initial_shooting':
-            // Executar rajada inicial
-            this.executeShootingBehavior(currentTime);
-            
-            // Verificar se completou o tiro inicial
-            if (this.burstCount >= this.maxBursts && !this.hasCompletedInitialShooting) {
-              this.hasCompletedInitialShooting = true;
-              this.setNextTarget();
-              this.state = 'moving_to_position';
-            }
-            break;
-            
-          case 'moving_to_position':
-            this.move();
-            break;
-            
-          case 'position_shooting':
-            // Executar rajada na posição
-            this.executeShootingBehavior(currentTime);
-            
-            // Verificar se completou o tiro na posição
-            if (this.burstCount >= this.maxBursts) {
-              // Boss nunca vai embora, sempre vai para próxima posição
-              this.setNextTarget();
-              this.state = 'moving_to_position';
-            }
-            break;
-        }
-      }
-      
-      private executeShootingBehavior(currentTime: number) {
-        if (!this.isInBurst) {
-          // Verificar se é hora de iniciar uma nova rajada
-          if (currentTime - this.fireTimer > this.fireRate && this.burstCount < this.maxBursts) {
-            this.isInBurst = true;
-            this.currentBurstShots = 0;
-            this.burstTimer = currentTime;
-          } else if (this.burstCount >= this.maxBursts) {
-            // Completou o tiro na posição, mover para próxima
-            this.setNextTarget();
-            this.state = 'moving_to_position';
-          }
-        } else {
-          // Executando rajada
-          if (currentTime - this.burstTimer > this.burstRate && this.currentBurstShots < this.shotsPerBurst) {
-            this.shoot();
-            this.currentBurstShots++;
-            this.burstTimer = currentTime;
-            
-            // Verificar se rajada terminou
-            if (this.currentBurstShots >= this.shotsPerBurst) {
-              this.isInBurst = false;
-              this.burstCount++;
-              this.fireTimer = currentTime;
-            }
-          }
-        }
-      }
-    }
-
-    // =============================================
-    // BOSS TYPE C
-    // =============================================
-    class BossTypeC extends Enemy {
-      // Estados do chefe (baseado no Enemy Type C)
-      private state: 'entering' | 'moving_to_position' | 'shooting' | 'moving' | 'leaving' = 'entering';
-      
-      // Sistema de disparo
-      private fireTimer = 0;
-      private fireRate = 300; // 800ms entre tiros
-      private currentShots = 0;
-      private shotsPerBurstC = 5; // 5 tiros por rajada
-      
-      // Sistema de movimento
-      protected moveSpeed = 180; // Velocidade aumentada
-      protected targetX = 0;
-      protected targetY = 0;
-      protected movementCountC = 0;
-      protected maxMovementsPerCycle = 3; // 3 movimentos por ciclo
-      private cycleCount = 0;
-      private maxCycles = 3; // 3 ciclos antes de sair
-      
-      // Sistema de dano visual
-      private isFlashing = false;
-      
-      constructor(scene: Phaser.Scene, x: number, y: number) {
-        super(scene, x, y, 'boss-C-frame-1');
-        
-        // Stats específicos do Boss Type C
-        this.hp = 70;
-        this.maxHp = 140;
-        this.speed = 180;
-        this.maxSpeed = 250;
-        this.shotsPerBurst = 5; // 5 tiros por rajada
-        this.maxShotsPerBurst = 8;
-        this.totalShots = 30; // Será ajustado pela onda
-        this.maxTotalShots = 100;
-        this.movementsBeforeLeaving = 999; // Nunca vai embora
-        this.maxMovements = 999;
-        
-        // Configurar sprite
-        this.sprite.setScale(1.5); // Maior que inimigos normais
-        this.sprite.setData('enemyType', 'BossC');
-        this.sprite.setData('hp', this.hp);
-        this.sprite.setData('maxHp', this.maxHp);
-        
-        // Calcular pontuação baseada na onda
-        const gameScene = this.scene as GameScene;
-        this.points = 5000 * (gameScene.currentWave || 1);
-        
-        // Definir primeira posição aleatória
-        this.setNextPosition();
-        
-        console.log(`Boss Type C criado com ${this.hp} HP`);
-      }
-      
-      // Sobrescrever método para calcular pontos baseado na onda
-      adjustForWave(wave: number) {
-        // Pontuação baseada na onda
-        this.points = 5000 * wave;
-        
-        // Aumentar shots per burst baseado na onda
-        this.shotsPerBurstC = Math.min(this.maxShotsPerBurst, 5 + Math.floor(wave / 2));
-        
-        // Aumentar total shots baseado na onda
-        this.totalShots = Math.min(this.maxTotalShots, 30 + (wave * 10));
-        
-        // Aumentar velocidade ligeiramente baseado na onda
-        this.speed = Math.min(this.maxSpeed, 180 + (wave * 5));
-        this.moveSpeed = this.speed;
-        
-        console.log(`Boss C ajustado para onda ${wave}: ${this.shotsPerBurstC} tiros/rajada, ${this.points} pontos`);
-      }
-      
-      private setNextPosition() {
-        // Gerar posição aleatória inicial na área de jogo
-        const preferredX = Phaser.Math.Between(150, 650);
-        const preferredY = Phaser.Math.Between(120, 250);
-        
-        // Boss não precisa verificar posições ocupadas (ele tem prioridade)
-        this.targetX = preferredX;
-        this.targetY = preferredY;
-      }
-      
-      move() {
-        const deltaTime = this.scene.game.loop.delta / 1000;
-        
-        // Calcular direção para o alvo
-        const directionX = this.targetX - this.sprite.x;
-        const directionY = this.targetY - this.sprite.y;
-        const distance = Math.sqrt(directionX * directionX + directionY * directionY);
-        
-        // Se chegou perto o suficiente do alvo
-        if (distance < 15) {
-          this.onReachedTarget();
-          return;
-        }
-        
-        // Normalizar direção e aplicar velocidade
-        const normalizedX = directionX / distance;
-        const normalizedY = directionY / distance;
-        
-        this.sprite.x += normalizedX * this.moveSpeed * deltaTime;
-        this.sprite.y += normalizedY * this.moveSpeed * deltaTime;
-        
-        // Atualizar animação de movimento
-        this.updateMovementAnimation(normalizedX);
-      }
-      
-      private onReachedTarget() {
-        switch (this.state) {
-          case "moving_to_position": {
-            // Chegou na posição inicial, começar a atirar e reservar posição
-            this.state = "shooting";
-            this.resetShooting();
-
-            const gameScene = this.scene as GameScene;
-            gameScene.reservePosition(this.sprite.x, this.sprite.y, this.id);
-            break;
-          }
-
-          case "moving": {
-            // Chegou na nova posição após completar tiros, começar a atirar novamente
-            this.state = "shooting";
-            this.resetShooting();
-
-            const gameScene = this.scene as GameScene;
-            gameScene.reservePosition(this.sprite.x, this.sprite.y, this.id);
-            break;
-          }
-
-          case "leaving":
-            // Saiu da tela
-            this.isDestroyed = true;
-            // Não marcar isKilledByPlayer como true - inimigo saiu da tela
-            break;
-        }
-      }
-      
-      shoot() {
-        // Reproduzir som do tiro do boss (volume maior para diferenciá-lo dos inimigos)
-        this.scene.sound.play('enemy-shoot', { volume: 0.4 });
-        
-        // Obter referência ao grupo de projéteis inimigos da scene
-        const gameScene = this.scene as GameScene;
-        if (!gameScene.enemyBullets) return;
-        
-        // Capturar posição atual do jogador no momento do disparo
-        if (!gameScene.player) return;
-
-        const playerX = gameScene.player.x;
-        const playerY = gameScene.player.y;
-        
-        // Calcular direção para o jogador
-        const directionX = playerX - this.sprite.x;
-        const directionY = playerY - this.sprite.y;
-        const distance = Math.sqrt(directionX * directionX + directionY * directionY);
-        
-        // Normalizar direção
-        const normalizedX = directionX / distance;
-        const normalizedY = directionY / distance;
-        
-        // Tentar criar projétil do boss
-        let bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + 40, 'boss-fire');
-        
-        // Sistema de fallback para garantir que o projétil seja criado
-        if (!bullet) {
-          // Limpar projéteis que saíram da tela para liberar espaço no grupo
-          gameScene.enemyBullets.children.entries.forEach((child: Phaser.GameObjects.GameObject) => {
-            const sprite = child as Phaser.GameObjects.Sprite;
-            if (sprite.active && (sprite.y > 650 || sprite.y < -50 || sprite.x < -50 || sprite.x > 850)) {
-              sprite.setActive(false).setVisible(false);
-            }
-          });
-          
-          // Tentar novamente após limpeza
-          bullet = gameScene.enemyBullets.get(this.sprite.x, this.sprite.y + 40, 'boss-fire');
-        }
-        
-        if (bullet) {
-          bullet.setActive(true);
-          bullet.setVisible(true);
-          bullet.setScale(1.2); // Projéteis do boss são maiores
-          
-          // Velocidade do projétil direcionado
-          const speed = 320;
-          const velocityX = normalizedX * speed;
-          const velocityY = normalizedY * speed;
-          
-          bullet.setData('velocityX', velocityX);
-          bullet.setData('velocityY', velocityY);
-          bullet.setData('damage', 2); // Boss causa mais dano
-          bullet.setData('isEnemyBullet', true);
-          bullet.setData('isBossBullet', true);
-          bullet.setData('isDirected', true); // Marcar como projétil direcionado
-        } else {
-          console.warn('Boss C não conseguiu criar projétil direcionado');
-        }
-        
-        this.currentShots++;
-        this.shotsFired++;
-      }
-      
-      private updateMovementAnimation(directionX: number) {
-        // Atualizar frame baseado na direção do movimento (apenas se não estiver em flash)
-        if (this.isFlashing) return;
-        
-        if (Math.abs(directionX) > 0.1) {
-          // Está se movendo horizontalmente - alternar entre frames 2 e 3
-          const frame = (this.scene.time.now % 600 < 300) ? 2 : 3;
-          const newTexture = `boss-C-frame-${frame}`;
-          
-          // Só mudar textura se for diferente da atual
-          if (this.sprite.texture.key !== newTexture) {
-            this.sprite.setTexture(newTexture);
-          }
-          
-          // Espelhar sprite baseado na direção
-          this.sprite.setFlipX(directionX > 0);
-        } else {
-          // Movimento apenas vertical ou parado - usar frame 1
-          if (this.sprite.texture.key !== 'boss-C-frame-1') {
-            this.sprite.setTexture('boss-C-frame-1');
-          }
-          this.sprite.setFlipX(false);
-        }
-      }
-      
-      private resetShooting() {
-        this.currentShots = 0;
-        this.fireTimer = this.scene.time.now;
-        
-        // Só mudar textura se não estiver em flash
-        if (!this.isFlashing && this.sprite.texture.key !== 'boss-C-frame-1') {
-          this.sprite.setTexture('boss-C-frame-1'); // Frame de repouso
-        }
-      }
-      
-      // Sobrescrever método takeDamage para adicionar efeito de flash
-      takeDamage(damage: number) {
-        this.hp -= damage;
-        console.log(`Boss C levou ${damage} de dano. HP restante: ${this.hp}/${this.maxHp}`);
-        
-        // Ativar efeito de flash
-        this.activateFlash();
-        
-        if (this.hp <= 0) {
-          this.destroy();
-        }
-      }
-      
-      private activateFlash() {
-        if (this.isFlashing) return; // Evitar múltiplos flashes simultâneos
-        
-        this.isFlashing = true;
-        console.log('Boss C piscando!');
-        
-        // Mudar para cor branca (flash)
-        this.sprite.setTintFill(0xffffff);
-        
-        // Voltar à cor normal após 250ms
-        this.scene.time.delayedCall(50, () => {
-          if (this.sprite && !this.isDestroyed) {
-            this.sprite.clearTint();
-            console.log('Flash do boss C terminado');
-          }
-          this.isFlashing = false;
-        });
-      }
-      
-      // Sobrescrever método destroy para múltiplas animações de destruição
-      destroy() {
-        if (this.isDestroyed) return;
-        
-        this.isDestroyed = true;
-        this.isKilledByPlayer = true;
-        
-        // Reproduzir som de morte do boss
-        this.scene.sound.play('boss-kill', { volume: 0.5 });
-        
-        console.log(`Boss Type C destruído! Pontuação: ${this.points}`);
-        
-        // Criar 9 animações de destruição: centro + 8 ao redor
-        const centerX = this.sprite.x;
-        const centerY = this.sprite.y;
-        const offset = 60; // Distância maior para boss maior
-        
-        // Animação central
-        new DeathAnimation(this.scene, centerX, centerY);
-        
-        // Animações ao redor (com pequeno delay para efeito visual)
-        this.scene.time.delayedCall(60, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY - offset); // Superior esquerdo
-          new DeathAnimation(this.scene, centerX, centerY - offset); // Superior centro
-          new DeathAnimation(this.scene, centerX + offset, centerY - offset); // Superior direito
-        });
-        
-        this.scene.time.delayedCall(120, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY); // Esquerda
-          new DeathAnimation(this.scene, centerX + offset, centerY); // Direita
-        });
-        
-        this.scene.time.delayedCall(180, () => {
-          new DeathAnimation(this.scene, centerX - offset, centerY + offset); // Inferior esquerdo
-          new DeathAnimation(this.scene, centerX, centerY + offset); // Inferior centro
-        });
-        
-        this.scene.time.delayedCall(240, () => {
-          new DeathAnimation(this.scene, centerX + offset, centerY + offset, () => {
-            // Callback final após todas as animações - iniciar transição de onda
-            const gameScene = this.scene as GameScene;
-            gameScene.onBossDefeated(this.points);
-          });
-        });
-        
-        // Remover sprite do boss
-        this.sprite.destroy();
-      }
-      
-      update() {
-        super.update();
-        if (this.isDestroyed) return;
-        
-        const currentTime = this.scene.time.now;
-        
-        // Máquina de estados
-        switch (this.state) {
-          case 'entering':
-            // Verificar se entrou na tela o suficiente
-            if (this.sprite.y > 80) {
-              this.state = 'moving_to_position';
-            } else {
-              // Continuar entrando (movimento para baixo)
-              this.sprite.y += this.moveSpeed * (this.scene.game.loop.delta / 1000);
-            }
-            break;
-            
-          case 'moving_to_position':
-          case 'moving':
-            this.move();
-            break;
-            
-          case 'shooting':
-            // Executar comportamento de tiro
-            if (currentTime - this.fireTimer > this.fireRate) {
-              if (this.currentShots < this.shotsPerBurstC) {
-                this.shoot();
-                this.fireTimer = currentTime;
-              } else {
-                // Completou os tiros, determinar próxima ação
-                this.onCompletedShooting();
-              }
-            }
-            break;
-        }
-      }
-      
-      private onCompletedShooting() {
-        this.cycleCount++;
-        
-        // Boss nunca vai embora, sempre continua em ciclos infinitos
-        // Reiniciar contador de ciclos quando atingir o máximo
-        if (this.cycleCount >= this.maxCycles) {
-          this.cycleCount = 0;
-        }
-        
-        // Sempre começar próximo ciclo de movimento
-        this.setNextPosition();
-        this.state = 'moving';
-        this.movementCountC = 0;
-      }
-    }
 
     // =============================================
     // CLASSE DA CENA DO JOGO
@@ -2355,8 +1190,8 @@ const GameCanvas: React.FC<GameCanvasProps> = ({ backgroundSpeed = .75, onNaviga
             const distance = Phaser.Math.Distance.Between(
               bulletSprite.x,
               bulletSprite.y,
-              this.currentBoss.sprite.x,
-              this.currentBoss.sprite.y
+              this.currentBoss.x,
+              this.currentBoss.y
             );
 
             if (distance < 50) {
@@ -2428,8 +1263,8 @@ const GameCanvas: React.FC<GameCanvasProps> = ({ backgroundSpeed = .75, onNaviga
             const distance = Phaser.Math.Distance.Between(
               this.player!.x,
               this.player!.y,
-              this.currentBoss.sprite.x,
-              this.currentBoss.sprite.y
+              this.currentBoss.x,
+              this.currentBoss.y
             );
 
             if (distance < 60) {
